+++ docs/changelog.md
@@ -10,7 +10,7 @@ A changelog:
 
 ## In progress
 
-* Only set the Content-Length header when necessary
+* Only set the body and Content-Length header when necessary
   ([#31](https://github.com/pyrates/roll/pull/31))
 
 ## 0.6.0 — 2017-11-22